Added correct namespace and renamed SetDiContainer
